[AIRFLOW-5782] Migrate AWS Lambda to /providers/amazon/aws [AIP-21] (#6518)
* [AIRFLOW-5782] Migrate AWS Lambda to /providers/aws [AIP-21]
