+++ airflow/contrib/hooks/aws_athena_hook.py
@@ -17,15 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""This module is deprecated. Please use `airflow.providers.aws.hooks.athena`."""
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.athena`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.aws.hooks.athena import AWSAthenaHook  # noqa
+from airflow.providers.amazon.aws.hooks.athena import AWSAthenaHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.aws.hooks.athena`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.athena`.",
     DeprecationWarning,
     stacklevel=2,
 )
