+++ airflow/providers/amazon/aws/hooks/s3.py
