+++ tests/providers/amazon/aws/sensors/test_athena.py
@@ -20,8 +20,8 @@
 import unittest
 
 from airflow import AirflowException
-from airflow.providers.aws.hooks.athena import AWSAthenaHook
-from airflow.providers.aws.sensors.athena import AthenaSensor
+from airflow.providers.amazon.aws.hooks.athena import AWSAthenaHook
+from airflow.providers.amazon.aws.sensors.athena import AthenaSensor
 from tests.compat import mock
